+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 0.13.3
-* adjust error logging from `SmartNotifier`
+* adjust handling unsupported channel in `smartNotifierHandler`
 
 # 0.13.2
 * pass original error information from `SmartNotifierValidator`